Temporary fix for nonexistent time extents
If time_coverage_start or time_coverage_end was not present, the get_status method would previously attempt to access an attribute from None, which would cause an unhandled exception, causing the global status to not be updated.  This commit introduces a temporary fix for functionality in the form of a None type check in case either attribute happens to be missing, in which case the start and/or end time are simply not calculated.